Added Basic Gallery
Added very simple gallery structure in form of a flex box. I will add a lightbox to it as well. For now the gallery will stay simple since retrieval of images will be done from BLOB image files saved as ASCII. Once the image saving on the server is done I will add full functionality to the gallery