Array of strokes is emitted along with userId
Changed strokes array emit to also emit socketId. Depending on the socketID the strokes array is either being pushed back to top, mid or bottom strokes array on the server. When a stroke array is pushed back, a function is being called to check if each array hold at least one stroke array. If that is the case then a complete drawing can be created.